refactor: 패키지 구조 변경 및 코드 전반 수정
