+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/domain/model/recruitment/Recruitment.java
@@ -1,6 +1,6 @@
-package com.trillionares.tryit.product.recruitment.domain.model;
+package com.trillionares.tryit.product.domain.model.recruitment;
 
-import com.trillionares.tryit.product.recruitment.domain.model.type.RecruitmentStatus;
+import com.trillionares.tryit.product.domain.model.recruitment.type.RecruitmentStatus;
 import jakarta.persistence.*;
 import lombok.*;
