+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/controller/CategoryController.java
@@ -1,9 +1,9 @@
 package com.trillionares.tryit.product.presentation.controller;
 
-import com.trillionares.tryit.product.domain.service.CategoryService;
+import com.trillionares.tryit.product.application.service.CategoryService;
 import com.trillionares.tryit.product.presentation.dto.BaseResponseDto;
-import com.trillionares.tryit.product.presentation.dto.CategoryIdResponseDto;
-import com.trillionares.tryit.product.presentation.dto.CategoryInfoRequestDto;
+import com.trillionares.tryit.product.presentation.dto.response.CategoryIdResponseDto;
+import com.trillionares.tryit.product.presentation.dto.request.CategoryInfoRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
