+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/controller/ProductController.java
@@ -3,11 +3,11 @@ package com.trillionares.tryit.product.presentation.controller;
 import com.querydsl.core.types.Predicate;
 import com.trillionares.tryit.product.domain.common.message.ProductMessage;
 import com.trillionares.tryit.product.domain.model.product.Product;
-import com.trillionares.tryit.product.domain.service.ProductService;
+import com.trillionares.tryit.product.application.service.ProductService;
 import com.trillionares.tryit.product.presentation.dto.BaseResponseDto;
-import com.trillionares.tryit.product.presentation.dto.ProductIdResponseDto;
-import com.trillionares.tryit.product.presentation.dto.ProductInfoRequestDto;
-import com.trillionares.tryit.product.presentation.dto.ProductInfoResponseDto;
+import com.trillionares.tryit.product.presentation.dto.response.ProductIdResponseDto;
+import com.trillionares.tryit.product.presentation.dto.request.ProductInfoRequestDto;
+import com.trillionares.tryit.product.presentation.dto.response.ProductInfoResponseDto;
 import com.trillionares.tryit.product.presentation.exception.CategoryNotFoundException;
 import com.trillionares.tryit.product.presentation.exception.ProductNotFoundException;
 import java.util.List;
