+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/dto/request/UpdateRecruitmentRequest.java
@@ -1,4 +1,4 @@
-package com.trillionares.tryit.product.recruitment.presentation.dto.request;
+package com.trillionares.tryit.product.presentation.dto.request;
 
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.Positive;
