+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/dto/request/UpdateRecruitmentStatusRequest.java
@@ -1,0 +1,8 @@
+package com.trillionares.tryit.product.presentation.dto.request;
+
+import com.trillionares.tryit.product.domain.model.recruitment.type.RecruitmentStatus;
+
+public record UpdateRecruitmentStatusRequest (
+        RecruitmentStatus status
+){
+}
