+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/dto/response/ProductInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.trillionares.tryit.product.presentation.dto;
+package com.trillionares.tryit.product.presentation.dto.response;
 
 import com.trillionares.tryit.product.domain.model.product.Product;
 import java.util.List;
