+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/dto/response/RecruitmentIdResponse.java
@@ -1,4 +1,4 @@
-package com.trillionares.tryit.product.recruitment.presentation.dto.response;
+package com.trillionares.tryit.product.presentation.dto.response;
 
 import java.util.UUID;
