+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/presentation/dto/response/UpdateRecruitmentStatusResponse.java
@@ -1,0 +1,10 @@
+package com.trillionares.tryit.product.presentation.dto.response;
+
+import com.trillionares.tryit.product.domain.model.recruitment.type.RecruitmentStatus;
+import java.util.UUID;
+
+public record UpdateRecruitmentStatusResponse(
+        UUID recruitmentId,
+        RecruitmentStatus status
+) {
+}
