--- com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/recruitment/presentation/dto/request/UpdateRecruitmentStatusRequest.java
@@ -1,8 +1,0 @@
-package com.trillionares.tryit.product.recruitment.presentation.dto.request;
-
-import com.trillionares.tryit.product.recruitment.domain.model.type.RecruitmentStatus;
-
-public record UpdateRecruitmentStatusRequest (
-        RecruitmentStatus status
-){
-}
